# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Driver Supervising Server API",
    "description": "Mandevices server",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "API cho phần người dùng"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/admin/signup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Tạo Admin mới trong csdl",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Thông tin về admin mới như trong model",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin mới đã được tạo",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          },
          "401": {
            "description": "Lỗi tạo admin"
          }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Đăng nhập admin",
        "parameters": [
          {
            "name": "admin",
            "in": "body",
            "description": "Thông tin về admin mới như trong model",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin của admin, kết quả trong user",
            "schema": {
              "$ref": "#/definitions/Admin"
            }
          },
          "401": {
            "description": "Lỗi đăng nhập"
          }
        }
      }
    },
    "/api/user/all": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Lấy thông tin tất cả user",
        "responses": {
          "200": {
            "description": "Thông tin của tất cả User, kết quả trong users",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "401": {
            "description": "Lỗi lấy thông tin user"
          }
        }
      }
    },
    "/api/user/add": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "user",
            "description": "Các thông tin về User",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "description": "Thêm một User vào database",
        "responses": {
          "200": {
            "description": "Thông tin của User mới, kết quả trong user",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Lỗi lấy thông tin user"
          }
        }
      }
    },
    "/api/data/add": {
      "post": {
        "tags": [
          "Data",
          "User"
        ]
      },
      "parameters": [
        {
          "name": "data",
          "description": "Thông tin về data mới, user là id của user sở hữu data này",
          "in": "body",
          "schema": {
            "$ref": "#/definitions/Data"
          }
        }
      ],
      "description": "Thêm Data mới ứng với người dùng vào csdl",
      "responses":{
        "200": {
          "description": "Data mới được thêm",
          "schema":{
            "$ref":"#/definitions/Data"
          }
        },
        "400": {
          "description": "Không thể thêm data"
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "fullName",
        "idCode",
        "phone",
        "dob",
        "certiType"
      ],
      "properties": {
        "fullName": {
          "type": "string"
        },
        "idCode": {
          "type": "string",
          "uniqueItems": true
        },
        "phone": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "certiType": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "learningTime": {
          "type": "number",
          "default": 0
        }
      }
    },
    "Admin": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Data": {
      "required": [
        "user",
        "lat",
        "long",
        "speed",
        "detected"
      ],
      "properties": {
        "user": {
          "type": "string"
        },
        "lat": {
          "type": "number"
        },
        "long": {
          "type": "number"
        },
        "speed": {
          "type": "number"
        },
        "detected": {
          "type": "number"
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    }
  }
}